artist pageのcreatePlaylistをモジュール化
