+++ lib/spotifyApiModule.js
@@ -2,32 +2,18 @@ import axios from 'axios';
 import config from '../config';
 
 // TODO accssToken, headersを共通化する
-module.exports.getUserId = async () => {
-  const accessToken = localStorage.getItem('accessToken');
-  const headers = { Authorization: `Bearer ${accessToken}` };
-
-  const endpoint = `${config.API_URL}/me`;
-  const user_id = await axios
-    .get(endpoint, { headers })
+module.exports.getUserId = async (headers) => {
+  return await axios
+    .get(`${config.API_URL}/me`, { headers })
     .then((res) => {
       return res.data.id;
     })
     .catch((error) => {
       throw error.response;
     });
-    return user_id;
 }
 
-module.exports.getPlaylistId = async (user_id) => {
-  const accessToken = localStorage.getItem('accessToken');
-  const headers = { Authorization: `Bearer ${accessToken}` };
-  
-  const playlistsConfig = {
-    name: 'Playlists of your favorite tracks',
-    description: 'Playlists of your favorite tracks',
-    public: true,
-  };
-
+module.exports.getPlaylistId = async (headers, playlistsConfig, user_id) => {
   return await axios
     .post(`${config.API_URL}/users/${user_id}/playlists`, playlistsConfig, {
       headers,
@@ -40,10 +26,21 @@ module.exports.getPlaylistId = async (user_id) => {
     });
 }
 
-module.exports.createPlaylist = async (playlistId, tracks_uri) => {
-  const accessToken = localStorage.getItem('accessToken');
-  const headers = { Authorization: `Bearer ${accessToken}` };
+module.exports.getArtistTrackUris = async (headers, artists) => {
+  return await Promise.all(
+    artists.map(async (artist) => {
+      const topTrackEndpoint = `${config.API_URL}/artists/${artist.id}/top-tracks?market=JP`;
+      return await axios
+        .get(topTrackEndpoint, { headers })
+        .then((res) => res.data.tracks[0].uri)
+        .catch((error) => {
+          throw error.response;
+        });
+    })
+  );
+}
 
+module.exports.createPlaylist = async (headers, playlistId, tracks_uri) => {
   return await axios
     .post(
       `${config.API_URL}/playlists/${playlistId}/tracks`,
